# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Services Project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "List all customers",
                "operationId": "2b23ac696cfd81c23807c2c1fd63b0b0",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current Page",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Order: latest or oldest",
                        "schema": {
                            "default": "latest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "create new customer",
                "operationId": "428d5533c8386f91d2baf4de69f01468",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "customer name",
                                        "default": ""
                                    },
                                    "email": {
                                        "description": "customer email",
                                        "default": ""
                                    },
                                    "phone": {
                                        "description": "customer phone",
                                        "default": ""
                                    },
                                    "address": {
                                        "description": "customer address",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created"
                    },
                    "422": {
                        "description": "validation error"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/customers/{customer_id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "List all customers",
                "operationId": "42bccb08ecb6db9ca22b872566de11ac",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "show single customer",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "update customer",
                "operationId": "74902da441a95a5c4359805841436f58",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "customer id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "customer name",
                                        "default": ""
                                    },
                                    "email": {
                                        "description": "customer email",
                                        "default": ""
                                    },
                                    "phone": {
                                        "description": "customer phone",
                                        "default": ""
                                    },
                                    "address": {
                                        "description": "customer address",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "422": {
                        "description": "validation error"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "delete customer",
                "operationId": "36fa87c70e3615b4a19340e254c9615d",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "description": "customer id",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "deleted"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "should send the Authorization header with \"Bearer <jwt_token>\"",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}